fix: prevented replacing existing symbolic links

+++ scripts/ubuntu/setup_cmdtools.sh
@@ -10,10 +10,18 @@ rm  ripgrep_14.1.0-1_amd64.deb
 
 # bat
 sudo apt install -y bat
-mkdir -p ~/.local/bin
-ln -s $(which batcat) ~/.local/bin/bat
+if [ -e ~/.local/bin/bat ] ; then
+  echo "fd already exists"
+else
+  mkdir -p ~/.local/bin
+  ln -s $(which batcat) ~/.local/bin/bat
+fi
 
 # fd
 sudo apt install -y fd-find
-mkdir -p ~/.local/bin
-ln -s $(which fdfind) ~/.local/bin/fd
+if [ -e ~/.local/bin/fd ] ; then
+  echo "fd already exists"
+else
+  mkdir -p ~/.local/bin
+  ln -s $(which fdfind) ~/.local/bin/fd
+fi
